requirements.txt
#system packages:
pulseaudio
pactl
uptime
date
python3

#python packages
python-mpd2
